190321 delete 완료 update 작성 중
성수형 서블릿 중에 mapping이 /toriginup 가 있어 충돌이 남.
내일 수정 요망.

+++ animalProject/web/views/tipboard/tipBoardDetailView.jsp
@@ -10,7 +10,7 @@
 <html>
 <head>
 <meta charset="UTF-8">
-<title>Insert title here</title>
+<title>팁게시판</title>
 <link href="/doggybeta/resources/css/footer.css" rel="stylesheet" type="text/css">
 </head>
 <body>
@@ -53,7 +53,7 @@
 	<%-- <% if(loginUser.getUserId().equals(board.getBoardWriter())){ %> --%>
 		<a href="/doggybeta/tupview?tnum=<%=tboard.getTipBoardNo()%>&page=<%=currentPage%>">[수정페이지로 이동]</a>
 		&nbsp; &nbsp;
-		<a href="/doggybeta/tdelete?tnum=<%=tboard.getTipBoardNo()%>">[글삭제]</a>
+		<a onclick="return confirm('정말로 삭제하시겠습니까?')" href="/doggybeta/tdelete?tnum=<%=tboard.getTipBoardNo()%>">[글삭제]</a>
 	<%-- <% } %> --%>
 	&nbsp; &nbsp;
 	<a href="/doggybeta/tlist?page=<%= currentPage%>">[목록]</a>
